Modifiche varie

+++ HomeActivity/app/src/main/java/Server/Server.java
@@ -247,10 +247,9 @@ public class Server {
     }
 
 
-
     //Metodo per le chiamate dell'API di Google Maps per il reverse geocoding
     //Problemi di conversione double a String
-    public static void callReverseGeocoding(Context context, final double[] coordinate, final VolleyCallback volleyCallback)
+    public static void callReverseGeocoding(Context context, final float[] coordinate, final VolleyCallback volleyCallback)
     {
         String indirizzo_p1 ="https://maps.googleapis.com/maps/api/geocode/json?latlng=";
         String indirizzo_p2="&key=AIzaSyA09MsWgTgGZvMh8KDjyNtxm8ovRYoq1Dg";
